revised HealthIndicatorTest
